implemented crud_open, and added a few helper functions
*NOTE: IMPORTANT CHANGE*
Rather than assigning a position value within processRequest where we
used to, we now must assign the returned value a position value. This is
because we are nolonger maintaining a single open file, but instead a
list of files and to get the position value we would have to know the
file handle of the file we're currently manipulating, which
processResponse() does not. We could pass it as a parameter, but I've
decided against that since I would just be doing basically the same
thing, and I think it's better to show how we're assigning that value in
each instance.